+++ czkawka_cli/AGENTS.md
@@ -159,3 +159,11 @@ and stop gracefully when it is `true`.
 - `libraw` – RAW image support (requires libraw)
 - `libavif` – AVIF image support (requires libavif)
 - `xdg_portal_trash` – FlatPak-compatible trash (XDG portal)
+
+---
+
+## Comments
+
+Short and minimal - see root `AGENTS.md`. Only add a comment where the code's behavior can't be
+inferred from reading it (a non-obvious constraint, a workaround, a cross-module coupling);
+never restate what the code already says.
