+++ czkawka_core/src/common/process_utils.rs
@@ -1,13 +1,13 @@
-use std::collections::HashMap;
 use std::process::{Command, Stdio};
 use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::{Arc, Mutex, OnceLock};
+use std::sync::{Arc, Mutex};
 use std::thread;
-use std::time::{Duration, Instant};
+use std::time::Duration;
 
-use log::{error, warn};
+use log::error;
 
 use crate::flc;
+use crate::helpers::long_operation_watcher::LongOperationWatcher;
 
 #[cfg_attr(not(target_os = "windows"), expect(clippy::needless_pass_by_ref_mut))]
 pub fn disable_windows_console_window(command: &mut Command) {
@@ -27,103 +27,6 @@ pub struct CommandOutput {
     pub status: std::process::ExitStatus,
     pub stdout: String,
     pub stderr: String,
-}
-
-// Seconds after which a still-running operation logs a warning - shared by the subprocess
-// watchdog (`run_command_interruptible`) and the in-process watchdog below.
-const WARNING_STEPS_SECS: [u64; 4] = [50, 250, 1250, 6000];
-
-// A single shared watcher thread that logs a warning when any registered in-process
-// operation runs past the WARNING_STEPS_SECS thresholds. Replaces the old "one watchdog
-// thread per file" approach: callers register an operation with `start`/`watch` and the
-// lone background thread, woken every 100 ms, scans the registry, removes nothing on its
-// own (entries leave via `stop`/the RAII guard) and emits escalating warnings so a hang on
-// a single file is visible in the logs instead of silently freezing the scan.
-struct OperationRecord {
-    operation: String,
-    start: Instant,
-    next_warning_idx: usize,
-}
-
-static WATCHER_OPERATIONS: OnceLock<Mutex<HashMap<String, OperationRecord>>> = OnceLock::new();
-static WATCHER_THREAD_STARTED: OnceLock<()> = OnceLock::new();
-
-fn watcher_operations() -> &'static Mutex<HashMap<String, OperationRecord>> {
-    WATCHER_OPERATIONS.get_or_init(|| Mutex::new(HashMap::new()))
-}
-
-fn ensure_watcher_thread() {
-    if WATCHER_THREAD_STARTED.set(()).is_err() {
-        return; // Already running - one thread for the whole process lifetime.
-    }
-    let spawn_result = thread::Builder::new().name("long-op-watcher".to_string()).spawn(|| {
-        loop {
-            thread::sleep(Duration::from_millis(100));
-            let mut guard = watcher_operations().lock().unwrap_or_else(std::sync::PoisonError::into_inner);
-            for (key, record) in guard.iter_mut() {
-                let elapsed_secs = record.start.elapsed().as_secs();
-                while let Some(warning_time) = WARNING_STEPS_SECS.get(record.next_warning_idx).copied() {
-                    if elapsed_secs < warning_time {
-                        break;
-                    }
-                    warn!("Operation \"{}\" is still running after {warning_time} seconds, for: {key}", record.operation);
-                    record.next_warning_idx += 1;
-                }
-            }
-        }
-    });
-    if let Err(e) = spawn_result {
-        error!("Failed to spawn long-operation watcher thread: {e}");
-    }
-}
-
-pub struct LongOperationWatcher;
-
-impl LongOperationWatcher {
-    // Register a running operation under `key` (use a stable, unique-at-a-time key such as the
-    // file path). Prefer `watch`/`run_with_long_operation_warnings` so `stop` cannot be missed.
-    pub fn start(operation: &str, key: &str) {
-        ensure_watcher_thread();
-        watcher_operations().lock().unwrap_or_else(std::sync::PoisonError::into_inner).insert(
-            key.to_string(),
-            OperationRecord {
-                operation: operation.to_string(),
-                start: Instant::now(),
-                next_warning_idx: 0,
-            },
-        );
-    }
-
-    pub fn stop(key: &str) {
-        if let Some(operations) = WATCHER_OPERATIONS.get() {
-            operations.lock().unwrap_or_else(std::sync::PoisonError::into_inner).remove(key);
-        }
-    }
-
-    // RAII variant: `stop` runs on drop, so it fires on every exit path (early return, `?`, panic unwind).
-    #[must_use]
-    pub fn watch(operation: &str, key: &str) -> OperationWatch {
-        Self::start(operation, key);
-        OperationWatch { key: key.to_string() }
-    }
-}
-
-pub struct OperationWatch {
-    key: String,
-}
-
-impl Drop for OperationWatch {
-    fn drop(&mut self) {
-        LongOperationWatcher::stop(&self.key);
-    }
-}
-
-// Convenience wrapper around the shared watcher: registers `key`, runs `f`, then deregisters
-// (even on panic). For blocking work that cannot be interrupted (e.g. in-process video hashing
-// or a blocking ffmpeg read).
-pub fn run_with_long_operation_warnings<T, F: FnOnce() -> T>(operation: &str, key: &str, f: F) -> T {
-    let _watch = LongOperationWatcher::watch(operation, key);
-    f()
 }
 
 // Remember - Ok returned by this function does not necessarily mean that the command executed successfully
@@ -176,23 +79,14 @@ pub fn run_command_interruptible(mut command: Command, stop_flag: &Arc<AtomicBoo
         }
     });
 
-    let start_time = Instant::now();
-    let warning_steps = WARNING_STEPS_SECS;
-    let mut next_warning_idx = 0;
+    // pid is unique only while the child is alive, which is exactly the watcher's lifetime here.
+    let _watch = LongOperationWatcher::watch(&format!("command {command:?}"), &child.id().to_string());
 
     loop {
         if stop_flag.load(Ordering::Relaxed) {
             let _ = child.kill();
             let _ = child.wait();
             break;
-        }
-
-        let elapsed_secs = start_time.elapsed().as_secs();
-        if let Some(warning_time) = warning_steps.get(next_warning_idx)
-            && elapsed_secs >= *warning_time
-        {
-            warn!("Command is still running after {warning_time} seconds, for command: {command:?}");
-            next_warning_idx += 1;
         }
 
         match child.try_wait() {
