+++ czkawka_core/src/helpers/mod.rs
@@ -4,4 +4,5 @@ pub mod audio_checker;
 pub mod debug_timer;
 pub mod delayed_sender;
 pub mod ffprobe;
+pub mod long_operation_watcher;
 pub mod messages;
