+++ czkawka_core/src/tools/similar_videos/core.rs
@@ -18,12 +18,12 @@ use crate::common::cache::{CACHE_VERSION, CACHE_VIDEO_VERSION, load_and_split_ca
 use crate::common::config_cache_path::get_config_cache_path;
 use crate::common::dir_traversal::{DirTraversalBuilder, DirTraversalResult, inode, take_1_per_inode};
 use crate::common::model::{ToolType, WorkContinueStatus};
-use crate::common::process_utils::run_with_long_operation_warnings;
 use crate::common::progress_data::{CacheLoadPhase, ProgressData, SimilarVideosMode, SimilarVideosStage, ToolStage};
 use crate::common::progress_stop_handler::{check_if_stop_received, prepare_thread_handler_common};
 use crate::common::tool_data::{CommonData, CommonToolData};
 use crate::common::video_utils::{VIDEO_THUMBNAILS_SUBFOLDER, VideoMetadata, generate_thumbnail};
 use crate::flc;
+use crate::helpers::long_operation_watcher::run_with_long_operation_warnings;
 use crate::tools::similar_videos::{SimilarVideos, SimilarVideosParameters, VideoAudioEntry, VideosEntry};
 
 impl SimilarVideos {
